lvm: split CGO backend by build tags
